+++ visualization/app/codeCharta/ui/codeMap/codeMap.render.service.ts
@@ -8,6 +8,30 @@ import { CodeMapArrowService } from "./codeMap.arrow.service"
 import { CodeMapNode, Node } from "../../codeCharta.model"
 import { StoreService } from "../../state/store.service"
 import { isDeltaState } from "../../model/files/files.helper"
+import { FileState } from "../../model/files/files"
+
+const ONE_MB = 1024 * 1024
+
+export enum MAP_RESOLUTION_SCALE {
+	SMALL_MAP = 1,
+	MEDIUM_MAP = 0.5,
+	BIG_MAP = 0.25
+}
+export function getMapResolutionScaleFactor(files: FileState[]) {
+	let totalFileSize = 0
+	for (const file of files) {
+		totalFileSize += file.file.fileMeta.exportedFileSize
+	}
+
+	switch (true) {
+		case totalFileSize >= 7 * ONE_MB:
+			return 0.25
+		case totalFileSize >= 2 * ONE_MB:
+			return 0.5
+		default:
+			return 1
+	}
+}
 
 export class CodeMapRenderService {
 	constructor(