+++ visualization/app/codeCharta/ui/codeMap/rendering/geometryGenerator.spec.ts
@@ -4,7 +4,7 @@ import { STATE, TEST_NODE_ROOT } from "../../../util/dataMocks"
 import { BuildResult, GeometryGenerator } from "./geometryGenerator"
 
 describe("geometryGenerator", () => {
-	let geomGen : GeometryGenerator
+	let geomGen: GeometryGenerator
 	let storeService: StoreService
 	let testNodes: Node[]
 
@@ -16,15 +16,17 @@ describe("geometryGenerator", () => {
 	const mockStoreService = () => {
 		STATE.dynamicSettings.heightMetric = "a" // set to a, since it is the delta defined in TEST_NODE_ROOT
 
-		storeService = jest.fn().mockReturnValue( {
-			getState : jest.fn().mockReturnValue(STATE)
+		storeService = jest.fn().mockReturnValue({
+			getState: jest.fn().mockReturnValue(STATE)
 		})()
 	}
 
 	const setTestNodes = () => {
-		const updatedNode = {heightDelta : -50} // delta has to be negative why is that ?
-		return [TEST_NODE_ROOT].map(node => { return {...node,...updatedNode} })
-	} 
+		const updatedNode = { heightDelta: -50 } // delta has to be negative why is that ?
+		return [TEST_NODE_ROOT].map(node => {
+			return { ...node, ...updatedNode }
+		})
+	}
 
 	const initData = () => {
 		geomGen = new GeometryGenerator()
@@ -32,17 +34,17 @@ describe("geometryGenerator", () => {
 	}
 
 	describe("addBuilding", () => {
-		let buildResult : BuildResult
-		const setFlattened = (isFlat : boolean) => {
+		let buildResult: BuildResult
+		const setFlattened = (isFlat: boolean) => {
 			testNodes.forEach(element => {
 				element.flat = isFlat
-			});
+			})
 		}
 
 		it("should add delta to height when not flattened", () => {
 			setFlattened(false)
 			buildResult = geomGen.build(testNodes, null, storeService.getState(), true)
-			
+
 			expect(testNodes[0].flat).toBeFalsy()
 			expect(buildResult.desc.buildings[0].boundingBox.max.y).toBe(58)
 		})
@@ -50,7 +52,7 @@ describe("geometryGenerator", () => {
 		it("should not add delta to height when flattened", () => {
 			setFlattened(true)
 			buildResult = geomGen.build(testNodes, null, storeService.getState(), true)
-			
+
 			expect(testNodes[0].flat).toBeTruthy()
 			expect(buildResult.desc.buildings[0].boundingBox.max.y).toBe(8)
 		})