+++ visualization/app/codeCharta/ui/codeMap/rendering/intermediateVertexData.ts
@@ -1,4 +1,12 @@
+import { Node } from "../../../codeCharta.model"
 import { Vector2, Vector3 } from "three"
+
+export interface SurfaceInformation {
+	node: Node
+	surfaceStartIndex: number
+	minPos: Vector3
+	maxPos: Vector3
+}
 
 export class IntermediateVertexData {
 	positions: Vector3[]
@@ -9,6 +17,7 @@ export class IntermediateVertexData {
 	deltas: number[]
 
 	indices: number[]
+	floorSurfaceInformation: SurfaceInformation[]
 
 	constructor() {
 		this.positions = new Array<Vector3>()
@@ -19,6 +28,7 @@ export class IntermediateVertexData {
 		this.deltas = new Array<number>()
 
 		this.indices = new Array<number>()
+		this.floorSurfaceInformation = new Array<SurfaceInformation>()
 	}
 
 	addVertex(pos: Vector3, normal: Vector3, uv: Vector2, color: string, subGeomIndex: number, delta: number) {
@@ -34,5 +44,17 @@ export class IntermediateVertexData {
 
 	addFace(index0: number, index1: number, index2: number) {
 		this.indices.push(index0, index1, index2)
+	}
+
+	saveFloorSurfaceInformation(node: Node, minPos: Vector3, maxPos: Vector3) {
+		const surfaceInformation: SurfaceInformation = {
+			node,
+			// The starting index of the vertices of the floors top surface must be known.
+			// It is used to render the floor surface with a different (label) texture
+			surfaceStartIndex: this.indices.length,
+			minPos,
+			maxPos
+		}
+		this.floorSurfaceInformation.push(surfaceInformation)
 	}
 }